requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.17.0
atpublic==5.0
attrs==24.2.0
billiard==4.2.1
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
databricks-sdk==0.36.0
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.3
dvc==3.55.2
dvc-data==3.16.6
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.16.1
Flask==3.0.3
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
grandalf==0.8
graphene==3.4
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gto==1.7.1
gunicorn==23.0.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.4.0
iterative-telemetry==0.0.9
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
lightgbm==4.5.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.17.0
mlflow-skinny==2.17.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
omegaconf==2.3.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.8
seaborn==0.13.2
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
sqltrie==0.11.1
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.13.2
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.16.0
zc.lockfile==3.0.post1
zipp==3.20.2
